Fixed a NaN issue

+++ product.h
@@ -1,15 +1,15 @@
 
 
 struct InterestRateSwap {
-    InterestRateSwap(double* pricing_points_, double* floating_schedule_, double* fixed_schedule_, double notional_, double K_, double expiry_, double dtau_) :
+    InterestRateSwap(float* pricing_points_, float* floating_schedule_, float* fixed_schedule_, float notional_, float K_, float expiry_, float dtau_) :
         pricing_points(pricing_points_), floating_schedule(floating_schedule_), fixed_schedule(fixed_schedule_), notional(notional_), K(K_), expiry(expiry_), dtau(dtau_)
     {}
 
-    double* pricing_points;
-    double* floating_schedule;
-    double* fixed_schedule;
-    double notional;
-    double K;
-    double dtau;
-    double expiry;
+    float* pricing_points;
+    float* floating_schedule;
+    float* fixed_schedule;
+    float notional;
+    float K;
+    float dtau;
+    float expiry;
 };
